apiServer.py renamed & divided
tested w/ DMH
